requirements.txt
discord~=2.3.2
python-dotenv~=1.0.0
pytz~=2023.3
tzlocal~=5.0.1
pytest
freezegun